Added improved ti 20pin smd adapter for beagle bone black.
The prevous adapter was connected to TRST which should be actually
connected to the global RST line. The beagle bone black has a pulldown
on the nTRST line keeping the JTAG circuit always in reset. Added a
pullup resistor on the nTRST line to keep the JTAG circuit out of reset.